+++ packages/datatrak-web/src/views/LandingPage/DraftSurveysSection/DraftSurveyTile.tsx
@@ -1,0 +1,98 @@
+import React from 'react';
+import styled from 'styled-components';
+import { ActionsMenu } from '@tupaia/ui-components';
+import { SurveyIcon } from '../../../components';
+import { DatatrakWebSurveyResponseDraftsRequest } from '@tupaia/types';
+import { useDeleteSurveyResponseDraft } from '../../../api';
+import { Tile } from '../../../components';
+
+type DraftSurvey = DatatrakWebSurveyResponseDraftsRequest.DraftSurveyResponse;
+
+const StyledActionsMenu = styled(ActionsMenu)`
+  position: absolute;
+  top: 50%;
+  right: 1rem;
+  transform: translateY(-50%);
+  z-index: 1;
+
+  ${({ theme }) => theme.breakpoints.down('sm')} {
+    display: none;
+  }
+`;
+
+const StyledTile = styled(Tile)`
+  padding-right: 4rem;
+
+  svg {
+    color: ${props => props.theme.palette.text.tertiary};
+  }
+
+  ${({ theme }) => theme.breakpoints.down('sm')} {
+    padding-right: 1rem;
+    flex: 1;
+    span {
+      flex: 1;
+    }
+    a {
+      width: 100%;
+    }
+  }
+`;
+const TooltipText = styled.p`
+  font-weight: normal;
+  margin-block: 0;
+  text-align: center;
+  text-wrap: balance;
+`;
+
+const Menu = ({ draftId }: { draftId: string }) => {
+  const { mutate: deleteDraft, isLoading } = useDeleteSurveyResponseDraft(draftId);
+
+  const actions = [
+    {
+      label: 'Delete',
+      action: () => {
+        if (isLoading) return;
+        deleteDraft();
+      },
+      toolTipTitle: 'Delete draft',
+    },
+  ];
+
+  return (
+    <StyledActionsMenu
+      anchorOrigin={{ vertical: 'bottom', horizontal: 'right' }}
+      transformOrigin={{ vertical: 'top', horizontal: 'right' }}
+      options={actions}
+      includesIcons
+    />
+  );
+};
+
+export const DraftSurveyTile = ({
+  id,
+  surveyName,
+  surveyCode,
+  countryCode,
+  entityName,
+  screenNumber,
+}: DraftSurvey) => {
+  const entityText = entityName ?? countryCode;
+  const tooltip = (
+    <>
+      <TooltipText>{surveyName}</TooltipText>
+      <TooltipText>{entityText}</TooltipText>
+    </>
+  );
+  return (
+    <StyledTile
+      heading={surveyName ?? 'Draft survey'}
+      leadingIcons={<SurveyIcon />}
+      tooltip={tooltip}
+      to={`/survey/${countryCode}/${surveyCode}/${screenNumber}?draftId=${id}`}
+    >
+      {entityText}
+      <Menu draftId={id} />
+    </StyledTile>
+  );
+};
