+++ packages/datatrak-web/src/views/LandingPage/DraftSurveysSection/index.ts
@@ -1,0 +1,1 @@
+export * from './DraftSurveysSection';
